设备协议代码修改提交

+++ tunnel-business/src/main/java/com/tunnel/business/service/dataInfo/impl/SdDevicesProtocolServiceImpl.java
@@ -28,7 +28,7 @@ public class SdDevicesProtocolServiceImpl implements ISdDevicesProtocolService {
      * @return 设备协议
      */
     @Override
-    public SdDevicesProtocol selectSdDevicesProtocolById(Integer id) {
+    public SdDevicesProtocol selectSdDevicesProtocolById(Long id) {
         return sdDevicesProtocolMapper.selectSdDevicesProtocolById(id);
     }
 
@@ -87,7 +87,7 @@ public class SdDevicesProtocolServiceImpl implements ISdDevicesProtocolService {
      * @return 结果
      */
     @Override
-    public int deleteSdDevicesProtocolByIds(Integer[] ids) {
+    public int deleteSdDevicesProtocolByIds(Long[] ids) {
         return sdDevicesProtocolMapper.deleteSdDevicesProtocolByIds(ids);
     }
 
@@ -98,7 +98,7 @@ public class SdDevicesProtocolServiceImpl implements ISdDevicesProtocolService {
      * @return 结果
      */
     @Override
-    public int deleteSdDevicesProtocolById(Integer id) {
+    public int deleteSdDevicesProtocolById(Long id) {
         return sdDevicesProtocolMapper.deleteSdDevicesProtocolById(id);
     }
